Suppression du timer qui permet en scrutation de repartir en mode init au bout d'un certain temps sans détection. Réception octet puissance des générateurs.

+++ F429HMIPlatform-bsp/Project/MicroEJ/inc/ImplementationProtocole.h
@@ -63,7 +63,7 @@ typedef union{
 
 
 typedef struct{
-	uint8_t PuissanceTable; //sur 7bits
+	uint16_t PuissanceTable; //sur 7bits
 	t_InfosRecette recette;
 	uint16_t TempsEnCours; //en secondes
 	uint16_t Temperature;
